Support for  typed values in LXPs rather than strings.
Former-commit-id: 411fa561e10e8639ff4f3511e7f92258dbe7ab26

+++ linkage/src/main/java/uk/ac/standrews/cs/digitising_scotland/linkage/lxp_records/SameAs.java
@@ -17,7 +17,7 @@ public class SameAs extends AbstractLXP {
     public static final String SECOND = "second";
     @LXP_SCALAR(type = LXPBaseType.STRING)
     public static final String RELATIONSHIP = "relationship";
-    @LXP_SCALAR(type = LXPBaseType.FLOAT)
+    @LXP_SCALAR(type = LXPBaseType.DOUBLE)
     public static final String CONFIDENCE = "confidence";
